# gateway/build/src/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "PayrollRecord": {
        "properties": {
          "id": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "period": {
            "type": "string"
          },
          "basicSalary": {
            "type": "number",
            "format": "double"
          },
          "allowances": {
            "type": "number",
            "format": "double"
          },
          "deductions": {
            "type": "number",
            "format": "double"
          },
          "netPay": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PROCESSED", "PAID"]
          }
        },
        "required": [
          "id",
          "employeeId",
          "period",
          "basicSalary",
          "allowances",
          "deductions",
          "netPay",
          "status"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "PayrollCreateRequest": {
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "period": {
            "type": "string"
          },
          "basicSalary": {
            "type": "number",
            "format": "double"
          },
          "allowances": {
            "type": "number",
            "format": "double"
          },
          "deductions": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["employeeId", "period", "basicSalary", "allowances", "deductions"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT authorization header using the Bearer scheme. Format: Authorization: Bearer {token}"
      }
    }
  },
  "info": {
    "title": "Safee Analytics API",
    "version": "1.0.0",
    "description": "API documentation for Safee Analytics SaaS platform",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "Safee Analytics",
      "email": "support@safee-analytics.com"
    }
  },
  "paths": {
    "/payroll": {
      "get": {
        "operationId": "GetPayrollRecords",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "records": {
                      "items": {
                        "$ref": "#/components/schemas/PayrollRecord"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "records"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HR & Payroll"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreatePayrollRecord",
        "responses": {
          "201": {
            "description": "Payroll record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayrollRecord"
                }
              }
            }
          }
        },
        "tags": ["HR & Payroll"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollCreateRequest"
              }
            }
          }
        }
      }
    },
    "/payroll/process": {
      "post": {
        "operationId": "ProcessPayroll",
        "responses": {
          "200": {
            "description": "Payroll processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HR & Payroll"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "period": {
                    "type": "string"
                  }
                },
                "required": ["period"],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}